backend-upload/requirements.txt
pyjwt
pymysql
geopy
Pillow
